Extract reusable `FeedStore` helpers to shared ext.
- Adds new group for specs files
