+++ EssentialFeedTests/Feed Cache/CodableFeedStoreTests.swift
@@ -9,7 +9,7 @@ import Foundation
 import EssentialFeed
 import XCTest
 
-class CodableFeedStoreTests: XCTestCase, FailableFeedStore {
+class CodableFeedStoreTests: XCTestCase, FailableFeedStoreSpecs {
     
     override func setUp() {
         super.setUp()
@@ -224,61 +224,5 @@ extension CodableFeedStoreTests {
         let localFeed = mockUniqueFeedWithLocalRep().localRepresentation
         let validTimestamp = Date().feedCacheTimestamp(for: .notExpired)
         return (localFeed, validTimestamp)
-    }
-    
-    // MARK: Test Assertion Helpers
-    private func expect(_ sut: FeedStore,
-                        toCompleteRetrievalWith expectedResult: RetrievedCachedFeedResult,
-                        file: StaticString = #filePath,
-                        line: UInt = #line) {
-        
-        let exp = expectation(description: "Wait for retrieval completion")
-        sut.retrieve { receivedResult in
-            switch (expectedResult, receivedResult) {
-                case (.empty, .empty), (.failure, .failure):
-                    break
-                case let (.found(expectedFeed, expectedTimestamp), .found(receivedFeed, receivedTimestamp)):
-                    XCTAssertEqual(expectedFeed, receivedFeed, file: file, line: line)
-                    XCTAssertEqual(expectedTimestamp, receivedTimestamp, file: file, line: line)
-                default:
-                    XCTFail("Expected retrieval with \(expectedResult), got \(receivedResult) instead", file: file, line: line)
-            }
-            exp.fulfill()
-        }
-        
-        wait(for: [exp], timeout: 1.0)
-    }
-    
-    private func expect(_ sut: FeedStore,
-                        toCompleteRetrievalTwiceWith expectedResult: RetrievedCachedFeedResult,
-                        file: StaticString = #filePath,
-                        line: UInt = #line) {
-        expect(sut, toCompleteRetrievalWith: expectedResult, file: file, line: line)
-        expect(sut, toCompleteRetrievalWith: expectedResult, file: file, line: line)
-    }
-    
-    @discardableResult
-    private func insert(_ cache: (feed: [LocalFeedImage], timestamp: Date), to sut: FeedStore) -> Error? {
-        let exp = expectation(description: "Wait for retrieval completion")
-        var insertionError: Error?
-        sut.insert(cache.feed, cache.timestamp) { error in
-            insertionError = error
-            exp.fulfill()
-        }
-        wait(for: [exp], timeout: 1.0)
-        return insertionError
-    }
-    
-    @discardableResult
-    private func deleteCache(from sut: FeedStore) -> Error? {
-        let exp = expectation(description: "Wait for deletion completion")
-        var deletionError: Error?
-        sut.deleteCachedFeed { error in
-            deletionError = error
-            exp.fulfill()
-        }
-        
-        wait(for: [exp], timeout: 4.0)
-        return deletionError
     }
 }
